cart fix

+++ customer/views.py
@@ -1622,23 +1622,24 @@ class PaymentResponseDebit(View):
 				closing_order.order_status = 'ORDER_CLOSED'
 				closing_order.save()
 
-			#customer booking temp data clearing
-			customer_cart = CustomerCart.objects.prefetch_related(Prefetch('cart_service',queryset=CartService.objects.filter(is_active=True).prefetch_related(Prefetch('cart_service_floor',queryset=CartServiceFloor.objects.all(),to_attr='cart_service_floors')),to_attr='cart_services'),Prefetch('cart_schedule',queryset=CartSchedule.objects.filter(is_active=True),to_attr='cart_schedules')).get(id=evaluation_id_encrypted)
-			for cart_service in customer_cart.cart_services:
-				if cart_service.cart_service_floors:
-					for floor in cart_service.cart_service_floors:
-						floor.delete()
+			if order_status == 'CUSTOMER_BOOKING' :
+				#customer booking temp data clearing
+				customer_cart = CustomerCart.objects.prefetch_related(Prefetch('cart_service',queryset=CartService.objects.filter(is_active=True).prefetch_related(Prefetch('cart_service_floor',queryset=CartServiceFloor.objects.all(),to_attr='cart_service_floors')),to_attr='cart_services'),Prefetch('cart_schedule',queryset=CartSchedule.objects.filter(is_active=True),to_attr='cart_schedules')).get(id=evaluation_id_encrypted)
+				for cart_service in customer_cart.cart_services:
+					if cart_service.cart_service_floors:
+						for floor in cart_service.cart_service_floors:
+							floor.delete()
 
-				cart_service.delete()	
+					cart_service.delete()	
 
-			for cart_schedule in customer_cart.cart_schedules:
-				cart_schedule.delete()
+				for cart_schedule in customer_cart.cart_schedules:
+					cart_schedule.delete()
 
-			customer_cart.is_scheduled = False
-			customer_cart.total_cost = 0
-			customer_cart.cart_discount = 0
-			customer_cart.final_cost = 0
-			customer_cart.save()
+				customer_cart.is_scheduled = False
+				customer_cart.total_cost = 0
+				customer_cart.cart_discount = 0
+				customer_cart.final_cost = 0
+				customer_cart.save()
 
 			#pay and book &&& others
 			# pay_and_book = request.POST.get('udf4')
@@ -1666,23 +1667,24 @@ class PaymentResponseDebit(View):
 			order.amount_paid     += amount_paid
 			order.save()
 
-			#customer booking temp data clearing
-			customer_cart = CustomerCart.objects.prefetch_related(Prefetch('cart_service',queryset=CartService.objects.filter(is_active=True).prefetch_related(Prefetch('cart_service_floor',queryset=CartServiceFloor.objects.all(),to_attr='cart_service_floors')),to_attr='cart_services'),Prefetch('cart_schedule',queryset=CartSchedule.objects.filter(is_active=True),to_attr='cart_schedules')).get(id=evaluation_id_encrypted)
-			for cart_service in customer_cart.cart_services:
-				if cart_service.cart_service_floors:
-					for floor in cart_service.cart_service_floors:
-						floor.delete()
-						
-				cart_service.delete()	
+			if order_status == 'CUSTOMER_BOOKING' :
+				#customer booking temp data clearing
+				customer_cart = CustomerCart.objects.prefetch_related(Prefetch('cart_service',queryset=CartService.objects.filter(is_active=True).prefetch_related(Prefetch('cart_service_floor',queryset=CartServiceFloor.objects.all(),to_attr='cart_service_floors')),to_attr='cart_services'),Prefetch('cart_schedule',queryset=CartSchedule.objects.filter(is_active=True),to_attr='cart_schedules')).get(id=evaluation_id_encrypted)
+				for cart_service in customer_cart.cart_services:
+					if cart_service.cart_service_floors:
+						for floor in cart_service.cart_service_floors:
+							floor.delete()
+							
+					cart_service.delete()	
 
-			for cart_schedule in customer_cart.cart_schedules:
-				cart_schedule.delete()
+				for cart_schedule in customer_cart.cart_schedules:
+					cart_schedule.delete()
 
-			customer_cart.is_scheduled = False
-			customer_cart.total_cost = 0
-			customer_cart.cart_discount = 0
-			customer_cart.final_cost = 0
-			customer_cart.save()
+				customer_cart.is_scheduled = False
+				customer_cart.total_cost = 0
+				customer_cart.cart_discount = 0
+				customer_cart.final_cost = 0
+				customer_cart.save()
 			
 			#pay and book &&& others
 			# pay_and_book = request.POST.get('udf4')
